Move HEVAG decoder to its own file
