+++ src/libvgmstream.vcxproj
@@ -424,6 +424,7 @@
     <ClCompile Include="coding\ogg_vorbis_decoder.c" />
     <ClCompile Include="coding\ogl_vorbis_decoder.c" />
     <ClCompile Include="coding\pcm_decoder.c" />
+    <ClCompile Include="coding\psv_decoder.c" />
     <ClCompile Include="coding\psx_decoder.c" />
     <ClCompile Include="coding\SASSC_decoder.c" />
     <ClCompile Include="coding\sdx2_decoder.c" />
